Move lock_and_call to certbot.test.util
